+++ src/etoolbox/datazip/core.py
@@ -341,13 +341,15 @@ class DataZip(ZipFile):
             return out_list
 
     def _read_df(self, name: str) -> pd.DataFrame | pd.Series:
-        out = pd.read_parquet(BytesIO(super().read(name + ".parquet"))).squeeze()
+        out = pd.read_parquet(BytesIO(super().read(name + ".parquet")))
         if name not in self._no_pqt_cols:
-            return out
+            return out.squeeze()
         cols, names = self._no_pqt_cols[name]
         if isinstance(names, (tuple, list)) and len(names) > 1:
-            return out.set_axis(pd.MultiIndex.from_tuples(cols, names=names), axis=1)
-        return out.set_axis(pd.Index(cols, name=names[0]), axis=1)
+            return out.set_axis(
+                pd.MultiIndex.from_tuples(cols, names=names), axis=1
+            ).squeeze()
+        return out.set_axis(pd.Index(cols, name=names[0]), axis=1).squeeze()
 
     def _recursive_write(self, name: str, data: dict | list | tuple) -> bool:
         if isinstance(data, dict):